+++ source2/Kernel/System/Hardware/Cosmos.Hardware/BlockDevice/Partition.cs
@@ -6,31 +6,25 @@ using System.Text;
 namespace Cosmos.Hardware.BlockDevice {
   public class Partition : BlockDevice {
     BlockDevice mHost;
-    //TODO:UInt64
-    UInt32 mStartingSector;
+    UInt64 mStartingSector;
 
-    //TODO:UInt64 x 2
-    public Partition(BlockDevice aHost, UInt32 aStartingSector, UInt32 aSectorCount) {
+    public Partition(BlockDevice aHost, UInt64 aStartingSector, UInt64 aSectorCount) {
       mHost = aHost;
       mStartingSector = aStartingSector;
       mBlockCount = aSectorCount;
-
       mBlockSize = mHost.BlockSize;
     }
 
-    public override void ReadBlock(UInt32 aBlockNo, UInt32 aBlockCount, byte[] aData) {
-      // TODO:UInt64
-      UInt32 xHostBlockNo = mStartingSector + aBlockNo;
+    public override void ReadBlock(UInt64 aBlockNo, UInt32 aBlockCount, byte[] aData) {
+      UInt64 xHostBlockNo = mStartingSector + aBlockNo;
       CheckBlockNo(xHostBlockNo, aBlockCount);
       mHost.ReadBlock(xHostBlockNo, aBlockCount, aData);
     }
 
-    public override void WriteBlock(UInt32 aBlockNo, UInt32 aBlockCount, byte[] aData) {
-      // TODO:UInt64
-      UInt32 xHostBlockNo = mStartingSector + aBlockNo;
+    public override void WriteBlock(UInt64 aBlockNo, UInt32 aBlockCount, byte[] aData) {
+      UInt64 xHostBlockNo = mStartingSector + aBlockNo;
       CheckBlockNo(xHostBlockNo, aBlockCount);
       mHost.WriteBlock(xHostBlockNo, aBlockCount, aData);
     }
-
   }
 }
